Fix buildbucket ID links on build details page.
Fix the self referential link displayed on a build details page when the build
does not have build numbers enabled and uses buildbucket IDs instead. In this
case the link needs a b prefixed to the buildbucket ID in the URL or it will
be interpretted as a build number and result in a 404.

BUG=959327

Change-Id: Iadb04da8956499ba55797d1756c2ce40204a6981
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1594659
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>