[bb] Add canary support
When printing a canary build, indicate that it is canary.

Add -canary and -nocanary flags to `bb add`.
Adding two flags instead of one with a value because vast majority
of the time this flag will be used to force canary.
-nocanary is added only for consistency.

R=tandrii@chromium.org

Change-Id: I6fb04b6dc1de83ab116f6850987abe7967b59db9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1602993
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>