[GCE] Try a new hostname when receiving 409 from instance creation
409 occurs when a hostname is being reused. Previously, we checked if
the instance existed in the same zone and assumed it was the instance
created by this service, but if the instance existed in a different zone
we assumed it was a true hostname collision. Instance creation is done
using an idempotent request ID, which prevents duplicate operations from
being created for the same instance. Therefore, the only way a 409 can
occur for an instance that does belong to this service is if the initial
request to create the instance and the subsequent request to check on
the status of the instance occurred so far apart that the request ID was
no longer a valid deduplication ID, triggering a 409. In practice this
is extremely unlikely since instance status is checked every minute but
request IDs are valid for 24 hours. Therefore, it's actually unsafe to
assume a same-named instance in the same zone belongs to this service.

Now, treat 409s the same as any other permanent failure: generate a new
hostname and try again. This logic assumes simply that a hostname
collision occurred.

Bug: 897355
Change-Id: Ie9ec96ee545dde6328c05d17cb7febf829d35b68
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1602040
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: smut <smut@google.com>