Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> VERY lgtm.

This doesn't do what I want. I want tasks to be deduped if there's a task with the same DeduplicationKey currently in the queue, but instead it looks at tasks that were ever in the queue. As implemented this causes tasks to be triggered once and then never again (DeduplicationKeys seem to be valid for at least one hour, so long after the task has exited the queue new tasks with the same key continue to be deduped).

I could dedupe by time (e.g. always gen the same dedupe key if called within the same minute), but this doesn't solve the problem I want.

What I want is: if there is a task to manipulate entity E already in the queue, no matter how old it is, don't enqueue another task to manipulate E, because it's unnecessary.

Patch-set: 1
Reviewer: Gerrit User 1003211 <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0