[logging] Add gkelogger.SeverityTracker.
It tracks maximum severity seen by a LogEntryWriter. It is needed to deduce an
overall logging severity of an HTTP request. For example, if the request handler
emitted a log line with Error level, the log entry for the request itself should
have Error level too. This mimics GAE behavior.

R=jchinlee@chromium.org, nodir@chromium.org
BUG=959427

Change-Id: I361bafc514ec16f1824751a9eeee651105109041
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1615875
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>