[server] Integrate client portion of auth stack.
This CLs makes auth.GetRPCTransport(...) and similar functions of the server
auth stack work on k8s, when there's a service account private key available and
path to it is passed via -service-account-json flag.

The implementation is based on the client-side auth stack. This gives "for free"
ability to run servers locally using developer's credentials (or using
LUCI_CONTEXT auth, as employed in the integration test).

Some bells and whistles are added to make this use case smoother:
  * Actionable error message if there's no cached credentials.
  * Ability to the server locally as some other service account (without having
    to download this account's key first), via "act as" IAM feature.
  * Token cache dir is configurable, so that ~/.config/chrome_infra/auth cache
    dir can be mounted into a container when running servers via docker locally.

R=jchinlee@chromium.org, nodir@chromium.org, tandrii@chromium.org
BUG=959427

Change-Id: I3635f47d0690fa5a99d0a9dd5ac53ff023ab3aa9