[lucicfg] Add a very weak notion of "key namespaces".
A key kind can be prefixed with '@' to indicate that it represents a namespace.
Currently it affects only how the node title is converted to a string ("ns:a/b"
instead of default "ns/a/b").

Namespaces will be eventually used when loading multiple project configs into
single instance of lucicfg process.

R=tandrii@chromium.org
BUG=966603

Change-Id: Ibf8572f0c73e151ecc7311610ea152cfdd1a447f