[lucicfg] Add support for external builders to Milo configs.
Such builders can be referred to as "<project>:<bucket>/<builder>", e.g.

luci.console_view_entry(
    console_view = 'external builders',
    builder = 'project:ci/builder',
    ...
)

The implementation is based on the assumption that in the future one lucicfg
project config can load another, and entries in one config can refer to entries
in another (if they actually exist). In this version, such external entires
are created "on the fly".

R=tandrii@chromium.org
BUG=966603

Change-Id: I690e2a18d7692e7878c3fd23f14454b7568bd9c0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1629989
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>