[archiver] Fix symlink direction when root is a symlink
This happens in particular on macOS, as TMPDIR points to a symlink. Calling
filepath.EvalSymlinks() leads to a different path, which means that
isolating from there will always inline the symlinks.

Refactor the test case to be more readable, using less closure.

Stop silencing logs in tests, which made diagnosing harder.

Change-Id: Ia401ba4c588470d42592fb93da5e6eabba2123c5