[archiver] Fix symlink direction when root is a symlink
This happens in particular on macOS, as TMPDIR points to a symlink. Calling
filepath.EvalSymlinks() leads to a different path, which means that
isolating from there will always inline the symlinks.

The reason it worked on the CI is that $TMPDIR is overridden by the Swarming bot
to a directory that is not a symlink.

Refactor the test case to be more readable, using less closure.

Stop silencing logs in tests, which made diagnosing harder.

Change-Id: Ia401ba4c588470d42592fb93da5e6eabba2123c5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1639257
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>