Update patch set 3
Patch Set 3:

> Thanks for the quick review!

You're welcome.

> It seems like `isolate` and `isolated` both have an archive option, but it's backed by a different implementation. Can you give me a little background about what each is used for? We only use `isolate batcharchive` currently.

We use isolate batcharchive, but we have users that use the isolated tool directly.
The isolate file format a GYP dialect, thus it's python raw data. In Go we load it with json5, which kinda works but is definitely not technically correct. It was designed in the pre-GN days around 2011 when we used GYP. The isolate files were originally manually written. GN is much more descriptive than GYP ever was, and the isolate files are nowadays generated.
We eventually want to get rid of the isolate format, and the isolate tool. 

https://crbug.com/907585
https://chromium.googlesource.com/infra/luci/luci-py.git/+/master/appengine/isolate/doc/client/Design.md
https://gyp.gsrc.io/

Patch-set: 3