Add "critical" to field mask for Build -> BuildSummary data.
Add the field mask for the "critical" field to the call to fetch Build
protos so that the value may be received and transfered to BuildSummary
instances. This was missed in https://crrev.com/c/1649521 which started
handling critical but failed to fetch it.

TEST=make test
BUG=None

Change-Id: I9ce5298203394b151ec086ce21541f051d710518
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1656177
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>