cq: allow duplicate project{} blocks in config for CrOS migration.
This is obviously not the cleanest implementation,
for instance one could have added a map inside project{} block to
map each ref_regexp to a luci_percentage. However, this code
will be thrown away anyway and the chosen approach makes it simpler to
specify in CrOS side, making it easier to read and reason about
which ref is migrated and to which extent.

R=nodir

Bug: 973896
Change-Id: I6c32c3af6e51b97980593be208f0e77e73393b06
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1658617
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Andrii Shyshkalov <tandrii@chromium.org>