[cipd] Use a lock file to protect .cipd/pkgs/<index> guts.
Otherwise concurrent "cipd ensure" calls may fail in a myriad of interesting
ways. Note that we do not lock .cipd guts entirely. It allows to install (unzip)
different packages concurrently.

The implementation is structured in a way that easily allows removing dependency
on a FS lock implementation when vendoring if unwanted.

R=tandrii@chromium.org
BUG=972203

Change-Id: I39886582fc77c934d57e1f5f1f8cf48266d3db66