[buildbucket] Add CIPD packages
Configure CIPD packages in a proto-text config.
One config is used for both prod and canary packages.

Also add support for negation to BuilderPredicate message.
This makes its usage simpler.

The builder lists are placed in service configs so that we don't
end up in a situation that whitelist is scattered across chops
and we need to chase clients.

Bug: 939368
Change-Id: Id1c54ea0afe855d34617c6a8154f8bdbd4ed72f1