Prevented symlink in subdir from being deleted by deployer.
This CL 1378599 fixed directory <-> symlink upgrades, however it
introduce a new behavior which doesn't look right. Imagine a package has
contents 'a/1', and it is deployed to 'sub/dir'. It is believed that the
change intended to ensure 'a' isn't a symlink, however the
implementation also delete any symlink as a part of 'sub/dir'.

Change-Id: I8bee5344dae66afd5205d8942ed15fabf86ef514