Prevented symlink in subdir from being deleted by deployer.
CL https://crrev.com/c/1378599 fixed directory <-> symlink upgrades,
however it introduce a new behavior which doesn't look right.
Imagine a package has contents 'a/1', and it is deployed to
'sub/dir'. It is believed that the CL intended to ensure 'a' isn't
a symlink, however the implementation also delete any symlink that
is a part of 'sub/dir'.

This CL prevented symlink that is a part of 'sub/dir' being deleted.

Change-Id: I8bee5344dae66afd5205d8942ed15fabf86ef514
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1677600
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Tyler Ouyang <tylero@microsoft.com>