[milo] Do not check that a builder is defined
Builder page should continue to work for a builder that was removed
because users may still want to see history of that builder.
This is more consistent with buildbucket SearchBuilds API which returns
build even if a builder does not exist.

Going forward, this would be enable "ephemeral" builders, essentially
defined in a build request.

The motivation for this change is to significantly speedup builder page,
because caching of defined builders recently broke.

Note that SearchBuilds still controls access.
If the user does not have access to a bucket, SearchBuilds returns
a NOT_FOUND error, as opposed to an empty response.

Bug: 970053
Change-Id: Ie248bf065aa146cdeda7a41f517d9ca0cc3d5715