[luciexe] Do not use authctx.Context for luciexe internals.
Use plain in-process auth.Authenticator instead.

authctx.Context is a huge overkill compared to auth.Authenticator. It launches
a bunch of HTTP servers, creates a bunch of files, etc. And it is not really
necessary: luciexe never exports env vars of the system auth context, so all
these local HTTP servers just sit idle.

More importantly, the existing code structure (where both 'task' and 'system'
contexts are setup in identical way) may give false impression that it is
somehow possible to run e.g. 'git' or 'gsutil' under the system account from
inside the user executable process tree. This is not possible, since we never
export required configs/env vars. It used to be possible in old 'kitchen' to
support 'git fetch' when bootstrapping recipes, but this feature was cut.

R=iannucci@chromium.org, nodir@chromium.org, tandrii@chromium.org

Change-Id: Ib8d9c948ed9a5f6493d1d3d0ae9ec148f0eff6b4