[auth] Attempt to make authctx.Context harder to misuse.
The problem with it is that to correctly launch a subprocess we need to export
LUCI_CONTEXT and a bunch of other env vars simultaneously, but export mechanisms
are very different: LUCI_CONTEXT changes travel through context.Context and env
var changes are explicit.

LUCI_CONTEXT modification was made in Launch(...) and was returned via modified
context.Context. It is very easy to ignore it. ExportIntoEnv tried to remedy
this (and make API simpler) by explicitly exporting LUCI_CONTEXT from the context
prepared in Launch, but this approach has a fatal flaw: it is easy to unknowingly
clobber exported LUCI_CONTEXT value (luciexe did just that).

In this CL we try to move all environment modifications (both context.Context
and environ) into single Export function. It is now slightly harder to misuse
(assuming people read function comments, which is a big assumption...).

R=iannucci@chromium.org, nodir@chromium.org, tandrii@chromium.org

Change-Id: Iecfa2f4586c7bb99d22bff87035799ebf9a940f2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1692001
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>