[luciexe] Replace devshell protocol with GCE metadata emulation.
It is enabled everywhere where devshell used to be enabled before. If devshell
is explicitly disabled on a builder, this builder most likely depends on
predeployed keys and/or GCE metadata for authenticating gsutil calls. If we
enabled GCE metadata emulation, we'll break such builder.

R=iannucci@chromium.org, nodir@chromium.org
BUG=912673

Change-Id: Ifa9acc226dbaa89c3ef2244b464cabf4953cbb84
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1693333
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>