Get rid of appengine/static/.
helloworld_standard was last place where it was used. `appengine/static/` is
a confusing bag of symlinks that still require non-trivial boilerplate to setup.
The amount of effort needed to use them is comparable to linking to rpcexplorer/
directly (all apps have been converted to do just that).

This also remove last reference to GAE go1.9 in luci-go.

R=iannucci@chromium.org, tandrii@chromium.org

Change-Id: Ieb46da5c50e9b0dae25a9010fb7478a150c8bec4