[sync/dispatcher/buffer] Better leased batch tracking and drop logic.
Removes the false pretense that Buffer is safe to use from multiple
goroutines. Methods like `Len()` are inherently dangerous in such an
environment anyhow (since there would be a race between `Len()` and any
action the caller might take based on that information).

Alters the internal heap to keep track of Batches which are leased out;
this means that we don't lose track of a Batch until the user calls
ACK/NACK on it.

R=tandrii@chromium.org

Change-Id: I7d5d8ab43bb145ad28ec64c59924b2f5349da3d5