Update patch set 2
Patch Set 2:

From email:

Thanks for the comments. As I'm working through addressing them, it would be really helpful for me to understand what final state we want to achieve with this specific CL.

To elaborate: my understanding was that this CL was to stand up /a/ pipeline for use in deploying /some/ pieces of LUCI results (as they get implemented), while we settle on a more finalized pipeline for deploying LUCI GKE apps in general. That's why this code is just the helloworld app under examples, no talking to Spanner, no setting up gRPC endpoints. Among other things, including GOOGLE_APPLICATION_CREDENTIALS as args is /not/ meant to be a final state, rather, to help confirm that the helloworld app, well, "hello"s, and using "stable" as a deployment tag is also not expected to be a final state.

The changes I've already made:
+ Simply strip out all the dev/prod duplications, and upload images to gcr.io/luci-test-results. (My intention for separating out the two image repositories was to keep images intended for production clearly and cleanly in one group, and images that could encompass testing and experimentation separated into another group, but we can talk about that later.)
+ Modify Dockerfile and Makefile to use distroless, etc.
+ Address nits :)

Changes that I'd like to clarify if you want to see in /this/ CL---there is also of course plenty of middle ground too:
+ Use https://godoc.org/go.chromium.org/luci/server/auth for authentication.
+ Use Cloud Build to build images.
+ Move some or all k8s configs and other operational details/configuration to a separate repo/other places.

I'm happy to address any and all of these points (and others I've missed); my question is regarding seeing them in this CL rather than incorporating them in later CLs.

From Vadim:

Other than Dockerfile change (original Dockerfile was just not correct) I personally don't really care what happens *now*. Whatever happens, we can adapt it later when we know precisely what to adapt it too. My comments are only information, as in "I think this is what we'll need to happen soon", not a call to action. I haven't worked out all the details yet to ask for a concrete changes.

Patch-set: 2
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>