Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> From email:
> 
> Thanks for the comments. As I'm working through addressing them, it would be really helpful for me to understand what final state we want to achieve with this specific CL.
> 
> To elaborate: my understanding was that this CL was to stand up /a/ pipeline for use in deploying /some/ pieces of LUCI results (as they get implemented), while we settle on a more finalized pipeline for deploying LUCI GKE apps in general. That's why this code is just the helloworld app under examples, no talking to Spanner, no setting up gRPC endpoints. Among other things, including GOOGLE_APPLICATION_CREDENTIALS as args is /not/ meant to be a final state, rather, to help confirm that the helloworld app, well, "hello"s, and using "stable" as a deployment tag is also not expected to be a final state.

Here some expectations from this CL, and a general principle for future CLs:

The CL should be closer to the long term solution (to the best of our knowledge today), as much as possible: for all aspects that are not blocked by something owned by someone else. Examples:

Blocked: a dedicated repository for deployments.yaml is not established yet, so until Vadim comes up with a concrete recommendation and guidelines, let's keep it here and optimize for cost (delivery time, simplicity). We will move it later (Unless I missed concrete guidelines)

Blocked: Vadim is still thinking about gke.py and related stuff, so let's keep the Makefile that we already have until Vadim comes up with a proper solution: gke.py VS makefile vs something else.

NOT Blocked: building the Go binary outside of Docker file. Let's fix this in this CL.

- 

> 
> The changes I've already made:
> + Simply strip out all the dev/prod duplications, and upload images to gcr.io/luci-test-results. (My intention for separating out the two image repositories was to keep images intended for production clearly and cleanly in one group, and images that could encompass testing and experimentation separated into another group, but we can talk about that later.)
> + Modify Dockerfile and Makefile to use distroless, etc.
> + Address nits :)

I don't see them. Please upload your changes.

> Changes that I'd like to clarify if you want to see in /this/ CL---there is also of course plenty of middle ground too:
> + Use https://godoc.org/go.chromium.org/luci/server/auth for authentication.

Authentication is a great exercise for CL splitting: it is a distinct aspect, and this CL would be still useful and still-sufficient without authentication (because the app does not require auth yet). This is a good time to think: "this part requires more discussion and can be extracted, so that I can land what we agree on here".

One more principle: let's not land things that will not be useful. Example: if we know for sure that we won't use the env variable, let's not land it even temporarily, because we essentially will remove it in the next CL.

> + Use Cloud Build to build images.

Vadim, can you clarify pros/cons?
 
+ do not have to install docker (once)
- have to upload Go binary to the cloud (every time)

> + Move some or all k8s configs and other operational details/configuration to a separate repo/other places.

See above.

> 
> I'm happy to address any and all of these points (and others I've missed); my question is regarding seeing them in this CL rather than incorporating them in later CLs.
> 
> From Vadim:
> 
> Other than Dockerfile change (original Dockerfile was just not correct) I personally don't really care what happens *now*. Whatever happens, we can adapt it later when we know precisely what to adapt it too. My comments are only information, as in "I think this is what we'll need to happen soon", not a call to action. I haven't worked out all the details yet to ask for a concrete changes.

Patch-set: 2
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>