[starlarkproto] Vendor google.golang.org/protobuf for now.
It is not officially released yet, its API may be unstable. Additionally it
makes some interesting choices (like forced randomization of text proto
encoding) that makes usage of text protos in lucicfg unnecessarily hard.

Hopefully, the randomization will become optional in the future (when more
people discover it and find it inconvenience). If it is still there in the
official release, we'll have to reimplement text encoding yourselves.

For now, our vendored google.golang.org/protobuf will be used as internal
implementation detail of new starlarkprotov2 package that implements
Starlark<->Proto interop in a "self contained" way.

R=iannucci@chromium.org
BUG=986972

Change-Id: I5dc1d274f8a7bc9239e6cf6e2852f1c193fb0a39