[lucicfg] Compare textproto files semantically in `validate`.
This means we ignore any formatting differences, comments, etc. when comparing
generated configs to what's on the disk right now. This will be useful to avoid
noisy diffs when we release a new lucicfg with new protobuf library which uses
slightly different format of textproto.

Note that we don't transform the structure of proto messages in any way, e.g. we
don't try to unroll Buildbucket mixins or anything like that.

R=iannucci@chromium.org
BUG=986972

Change-Id: Ic2f48197066230b26c39d2aff8e03203b4a482ad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1755131
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>