[lucicfg] Migration to use real proto path in load("@proto//...").
While aliases (e.g. "luci/buildbucket/project_config.proto" instead of
"go.chromium.org/luci/buildbucket/proto/project_config.proto") are nice,
ultimately they are lies and introduce unnecessary layer of indirection. They
will not be available in new descriptor-based proto API.

If the real proto paths are ugly and inconsistent (as they are), they should be
fixed to be non-ugly in the source of truth.

R=iannucci@chromium.org
BUG=986972

Change-Id: I401009da73fc3d1a20218ad4cd5caa86beae218c