[vpython] Use OpenCurrentProcessToken to obtain SID on Windows.
`user.Current()` does this for obtaining SID (which is the only field we
actually use), but also does some potentially heavyweight calls to
populate the rest of the User struct (including potentially chatting
with a slow domain controller over the network).

The new implementation here should (I think) only do local operations,
even for an AD-joined Windows machine. Since we only need to disambiguate
the caches between local users, this should be good-enough.

R=vadimsh@chromium.org

Change-Id: I3fbbe0c403d00519c0e1e4b2db3552db9ed00108