[swarming] spawn-tasks: add option to cancel extra tasks
Due to a swarming bug, spawn-tasks may sometimes create
extra tasks which are not returned in spawn task responses.

This change introduces optional handling:

1) Create a UUID, InvocationUUID
2) Spawn tasks with all with tag InvocationUUID
3) Count tasks by tag InvocationUUID
4) If actual count > reported spawned tasks, we've spawned
   additional tasks to be killed
4a) List tasks by tag InvocationUUID (I don't list them
    initially because this call is slower than count)
4b) Create a hashset of task IDs from spawn task responses
4c) For each listed task, check if its task ID is in 4b.
    If not, submit a cancel task request

Additionally, to begin to quantify the amount of duplication
on all instances using the Go code via BigQuery, we will
generate and tag a RPCUUID for every task request, regardless
of whether the -cancel-extra-tasks flag is set. This way,
a task request which spawned two tasks is known by checking
duplicate RPCUUIDs.

Bug: fuchsia:35386
Bug: 997221

Change-Id: I4a56f9431bfe1c3d7514f35d9dab33d438fd5738