[logdog/butlerlib] Make client typesafe and correctly work for subprocs.
This removes the weird `Stream` union type in favor of `io.WriteCloser`
and its alter-ego `DatagramStream`. This also removes the need for the
client to handle the odd `streamproto.Flags` type, in favor of
well-documented Option functions.

Adds a single well-known (and tested) in-memory FakeClient, a Loopback
client as well as a `null` protocol. There are now no external
implementations of Client (or its new `dialer` type).

As a bonus, this adds full testing for all client protocols, on all
platforms, including ensuring that `ForProcess` file descriptors actually
are the correct types and correctly work for subprocesses.

R=vadimsh@chromium.org

Change-Id: I1eeb5e775f4763fc00a3ed9031c4ff75fef92a86