[server] Refactor the initialization to be mostly in New(...).
It used to be in ListenAndServe(...). By moving it to New(...) it becomes
possible to use the fully initialized context to do stuff (e.g. use Datastore)
before starting the serving loop.

It will also be possible to modify the root server context used by the request
handlers before starting the loop.

R=iannucci@chromium.org
BUG=959427

Change-Id: I26d8c331d50b222777fa0116069550d554b19eb4