[tokenserver] Use target.TaskType for tsmonFlags.Target.TargetType instead of "task".
The type of the flag is target.Type. Although target.Type is defined as
a string, it should be treated as target.Type, not string. That's why
--ts-mon-target-type is registered via Flags.Var(), not Flags.StringVar().

If application users pass --ts-mon-target-type via the command line arguments,
then flag.Parse() is reponsible to invoke Type.Set() to map the string value
to the desinated Type variable. However, if application wants to set
the tsmonFlags.TargetType directly, then it should assign an instance of
target.Type to tsmonFlags.TargetType, because the type of tsmonFlags.TargetType
is target.Type, not string.

Logdog is a good example of setting the flag with an instance of target.Type.
: https://cs.chromium.org/chromium/infra/go/src/go.chromium.org/luci/logdog/server/service/service.go?l=296&rcl=aff73f3d92e4d25e4ac0d633e6beb9f51c3c847f

Change-Id: Ib98045465250d706f3e8d4d749d80d6aee8370de
R=vadimsh