[starlark] Make proto descriptor sets more composable.
This CL introduces DescriptorSet Starlark value which represents a bunch of
proto files. A DescriptorSet can depend on zero or more other descriptor sets.
Such dependencies can be shared.

For example, descriptor sets A and B may both depend on well-known protos,
specified via descriptor set WK. When registering A and B in the loader, WK will
be registered only once.

This feature is necessary for lucicfg packages to share proto definitions.
A package that exposes some protos will expose their DescriptorSet as part of
its public API, so that other package can depend on it in their DescriptorSets.

R=iannucci@chromium.org
BUG=986972

Change-Id: Ibe80c846b1d7aec0283e523234acdd5b1b1548b5