[swarming] Ensure all task subdirectories of output dir are created
If -output-dir is passed, then the client will download each task's
isolated outputs to $OUT_DIR/$TASK_ID, *provided* that the task has
isolated outputs. This is inconsistent behavior - and the swarming
recipe module currently implicitly relies on the $TASK_ID subdirectory
being created for each task, regardless of whether it had isolated
outputs. This change makes it so all of these subdirectories are
created regardless of whether there were isolated outputs; an empty
subdirectory is a reasonable convention for "nothing was isolated".

Change-Id: I76e4dad0e26e99b1e62f3d010113a44045319b17