Update patch set 6
Patch Set 6: Code-Review+1

One challenge with using multiple Archive in parallel is that the number of outgoing simultaneous RPCs is up to number of files times the limits.

On the other hand doing it serially isn't efficient either, as if one Archive gets a 10Gb file, it'll wait for it to complete before starting the next one.

So basically there's an abstraction layer problem, where the design of Archive should be changed. :/

In the meantime, I'd say to add a caveat to the function documentation about the concurrent network requests, but that will be good enough for now.

Patch-set: 6
Reviewer: Gerrit User 1000487 <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1