[tokenserver] Use target.TaskType for tsmonFlags.Target.TargetType instead of "task".
* Background
https://crrev.com/c/1771678 changes the type of Flags.TargetType
from target.Type to types.TargetType, and this CL has been written to fix
unit test failures caused by the type changes.

* Why
The type of the flag is target.Type. Although target.Type is defined as
a string, it should be treated as target.Type, not string. That's why
--ts-mon-target-type is registered via Flags.Var(), not Flags.StringVar().

If application users pass --ts-mon-target-type via the command line arguments,
then flag.Parse() is reponsible to invoke Type.Set() to map the string value
to the designated Type variable. However, if application wants to set
the tsmonFlags.TargetType directly, then it should assign an instance of
target.Type to tsmonFlags.TargetType, because the type of tsmonFlags.TargetType
is target.Type, not string.

Logdog is a good example of setting the flag with an instance of target.Type.
: https://cs.chromium.org/chromium/infra/go/src/go.chromium.org/luci/logdog/server/service/service.go?l=296&rcl=aff73f3d92e4d25e4ac0d633e6beb9f51c3c847f

Change-Id: Ib98045465250d706f3e8d4d749d80d6aee8370de
R=vadimsh
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1771677
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>