[tsmon] Add types.TargetType in MetricInfo
To add support for custom-target schema in tsmon, a number of CLs have been
written to make target, metric, store, and registry modules targe-type aware.
This CL adds types.TargetType, which represents the type of the intended
monitoring Target for a given metric.

To get an idea of how this target-type interface is used from user's
perspectives, please find the description and code changes of
crrev.com/c/1772858.


* Backward-incompatible changes introduced by this CL.
Note that this CL also changes the type of tsmon/target/flags:Flags.TargetType
from target.Type to types.TargetType. This type change causes a compile-time
error to applications that
- assume that target.Type is an alias for string, and
- assign a string value to tsmon.target.flags.TargetType directly.

However, applications that assigns an instance of target.Type to the flag
do not suffer from the issue, and the following CLs have been merged to fix
the applications that directly assigns a string value to the flag.
- https://crrev.com/c/1771162
- https://crrev.com/c/1771677

Change-Id: Ie9003d923a2bb3c76aa4e0d58b3331a8734ab733
Bug: 941798
R=junyer,vadimsh
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1771678
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>