[logdog/butlerlib/streamclient] Uniquify pipe name in windows test.
It's possible (but unlikely) that something is hanging onto the pipe
HANDLE for much longer than it ought to on the test bot (when the bot
runs many tests in parallel).

Tweak the test so that every invocation gets a system-unique pipe name
to avoid this possibility.

Also quintuple the timebomb fuse on windows; The entire test completes in
.5s on my windows VM, as opposed to .1s on my mac laptop.

R=maruel@chromium.org, vadimsh@chromium.org

Bug: 998936
Change-Id: Icebdd10347fd186644ef0d9da9b5830abaaf74d5