[logdog/butlerlib/streamclient] Uniquify pipe name in windows test.
It's possible (but unlikely) that something is hanging onto the pipe
HANDLE for much longer than it ought to on the test bot (when the bot
runs many tests in parallel).

Tweak the test so that every invocation gets a system-unique pipe name
to avoid this possibility.

Also quintuple the timebomb fuse on windows; The entire test completes in
.5s on my windows VM, as opposed to .1s on my mac laptop.

R=maruel@chromium.org, vadimsh@chromium.org

Bug: 998936
Change-Id: Icebdd10347fd186644ef0d9da9b5830abaaf74d5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1776434
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>