[tsmon] Add TargetType support in metric
This CL adds the following features.

* Metric factory functions with TargetType.
With the new Metric factory functions, Metric objects can be created with
a TargetType. With a Metric associated a TargetType, store will be able to
resolve the target object for a given Metric, based on the TargetType.

To avoid backward-incompatible changes, the existing Metric functions create
Metric objects with target.NilType. If a Metric has target.NilType, then
the Target object of the Metric will be
(1) the target object, installed in context, with the same type of
the instance returned by store.DefaultObject() or
(2) the returned object of store.DefaultObject()

* Metrics are registered with (metric_name, target_type)
In the current implementation, registry panics if more than one metric objects
are created with the same metric name. In this CL, metrics are registered
with (metric_name, target_type), which allows multiple metric objects created
with the same name, but different target types.

This is a useful feature for metric migration work with a new TargetType.
Sometimes, architectural changes are made to applications, and it becomes
necessary to change the model of the monitored targets in
a backward-incompatible way. For smooth migration, there is typically a grace
period where the existing metric data are reported with both old and new
schemas. Once it has been verified that the monitoring graphs with the two
schemas agree with each other, then the metric with the old schema is turned
off.

R=vadimsh,junyer
Bug: 941798
Change-Id: I83513474ffdf0dd418c5a7a4d710666a57cb814e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1774157
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>