[butler/output] Add URLConstructionEnv to let Output do its own layer violation.
Since the majority of logdog envvars are really about letting clients
construct (currently incorrect) viewer URLs to streams, and the viewer URL
is really an artifact of the Output, not of the Butler, add a method
to the Output to let it barf out the related environment variables for
however it's actually configured.

Note that most of the Output implementations have minimal or no related
values.

I don't want to dig into this rabbit hole much more, but the URL construction
really should be an RPC to the butler so that there can be exactly one
environment variable instead of 5. Even if we stick to environment variables,
it should probably just be a single envvar like LOGDOG_VIEWER_URL_FORMAT_STRING
that looks like e.g. "https://some.host.org/or/whatever/$STREAM_NAME", since
"stream name" is the ONLY thing that the user environment actually cares
about.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 939368
Change-Id: I81702c002f31885499aad6fcea2981a5aac13636