[vpython] Expose mechanism to set default vpython spec
This is the spec to fall back on when no other spec can be found. In
particular, the spec to be used when a no-arg "vpython" invocation is
made. By exposing a mechanism to override the spec, we can e.g. have
"vpython3" set Python 3 as the preferred version when invoked with no
arguments and no spec file.

Bug: 898348
Change-Id: Iad5dc8e3a19a2a220aae2d5c94985893fe6300ae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1779158
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: smut <smut@google.com>