[tsmon] check if targetTypeFlag has been set.
If a program, with tsmon, starts with --help, then flagset attempts to invoke
String() of all the registered flags, and, as a result, targetTypeFlag.String()
is invoked.

However, the default value of the flag is targetTypeFlag{flags: nil}, and,
therefore, targetTypeFlags.String() panics with the following message.
: invalid memory access or deference to nil.

This CL fixes the bug by checking if flags is nil.

Bug: 941798
Change-Id: Ib3ec392b557bd5da73e3421ca37bae180c774f2b