[tsmon] check if targetTypeFlag has been set.
If a program starts with --help or invalid options, flagset iterates through
all of the registered flags and invokes String() of each of the registered flags
to render the default value of them.

The default value of targetTypeFlag is targetTypeFlag{flags: nil}, and
targetTypeFlag.String() is invoked in the above cases. However, it panics
with "invalid memory access or dereference to nil", because ttf.flags is nil in
the default value of the flag.

This CL fixes the bug by checking if ttf.flags is nil. If it is nil, then return
the name of the DeviceTarget, which is the default target type.

R=junyer@google.com
Bug: 941798
Change-Id: Ib3ec392b557bd5da73e3421ca37bae180c774f2b