[tsmon] add examples/beep.
examples/beep contains a demo program showing how to define a new monitoring
target type and metrics with the new type of targets.

If the program gets executed, it reports monitoring data with 4 different
DummyProject targerts, and prints out the generated MetricsCollection data
in stderr.

reviewers@,
the output of a sample run can be found at http://shortn/_MIq8hCGUhE

Bug: 941798
Change-Id: I3a9049b06f78e3bbb6556cd362bba07219272f90