Update patch set 3
Patch Set 3:

One general question: I talked with isamsonov@ and learnt that he has a Q4 OKR to switch gtests to use json-test-result-format and expectations, though priority and timeline TBD.

With this in mind, I'm wondering should we prioritize supporting json-test-result-format and continue with the rest of tasks to implement results db? And if gtests has not switched to json-test-results-format when everything else is ready, we then add the support to it?

Assuming supporting gtests does not block anything else.

Patch-set: 3
Reviewer: Gerrit User 1157909 <1157909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>